Implemented dice rolls for resources
Dice rolls for resources now work, but the thief is not yet implemented. Resources are not currently displayed, so testing is difficult (will add soon). Additionally, there seems to be a bug where you can build multiple different buildings at once (i.e. a road and a city) (the other buttons are supposed to be disabled while building).