Started work on development cards, and fixed some bugs.
The implementation of development cards is in progress. They're being implemented using an abstract method in the Board.DEVELOPMENT enum that is overridden based on the effect of a given card.

A bug was found involving JPopupMenus with the thief and trading features where the menu wouldn't always properly detect inputs. To fix this, a "helper class" was created: ForcedPopup. This class extends JPopupMenu, and prevents the popup from closing by itself (when the user clicks somewhere). Now, the user is forced to give it an input, only that input can let the popup close itself.