Fixed bug with longest road, and undid testing changes.
Longest road was being calculated improperly, since the graphs are not inherently acyclic (you're not allowed to go in loops, but there are still loops in the graph). To fix this, DFS was used instead of BFS.

Additionally, when ensuring that longestRoad was calculated from both starting settlements, the wrong corner was being used under the assumption that the second starting settlement would be the second "accessible corner" (it's actually the third one since the second one is added when the road is built from the first starting settlement).